lnwire: add RemoteSigsFirst TLV to OpenChannel2
This TLV was proposed by @t-bast in
https://github.com/niftynei/lightning-rfc/pull/9/files.  It is not
merged into the dual funding spec yet, but since the PR eliminates a DoS
attack vector, I expect it to be merged.

By default, the initiator of the dual fund sends their funding
transaction signatures first.  But when RemoteSigsFirst is present in
the OpenChannel2 message, the non-initiator needs to send their
signatures first.  This enables the initiator to batch multiple channel
opens in the funding transaction if the non-initiator obliges.

The TLV is always zero-length, since no additional information is needed
beyond the TLV's presence or absence.